content/media/test/test_bug895091.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=895091
-->
<head>
  <meta charset='utf-8'>
  <title>Test for Bug 895091 - Integrating vtt.js</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPrefEnv({"set": [["media.webvtt.enabled", true]]},
  function() {
    var video = document.createElement("video");
    video.src = "seek.webm";
    video.preload = "auto";

    var trackElement = document.createElement("track");
    trackElement.src = "long.vtt";
    trackElement.kind = "subtitles";

    var trackElementTwo = document.createElement("track");
    trackElementTwo.src = "long.vtt";
    trackElementTwo.kind = "subtitles";

    document.getElementById("content").appendChild(video);
    video.appendChild(trackElement);
    video.appendChild(trackElementTwo);

    video.addEventListener("loadedmetadata", function run_tests() {
      // Re-que run_tests() at the end of the event loop until the track
      // element has loaded its data.
      if (trackElement.readyState == 1 || trackElementTwo.readyState == 1) {
        setTimeout(run_tests, 0);
        return;
      }

      is(trackElement.readyState, 2, "Track::ReadyState should be set to LOADED.");
      is(trackElementTwo.readyState, 2, "Track::ReadyState should be set to LOADED.");

      is(trackElement.track.cues.length, 2000, "Cue list length should be 2000.");
      is(trackElementTwo.track.cues.length, 2000, "Cue list length should be 2000.");

      SimpleTest.finish();
    });
  }
);

</script>
</pre>
</body>
</html>
